refactor: rename managed auth utils module
