+++ packages/server/lib/controllers/v1/account/managed/auth.ts
