+++ packages/server/lib/controllers/v1/account/managed/getCallback.ts
@@ -3,7 +3,7 @@ import * as z from 'zod';
 
 import { basePublicUrl, getLogger } from '@nangohq/utils';
 
-import { finalizeManagedAuthentication, getManagedAuthEmailVerificationFromError, setManagedAuthEmailVerification } from './utils.js';
+import { finalizeManagedAuthentication, getManagedAuthEmailVerificationFromError, setManagedAuthEmailVerification } from './auth.js';
 import { getWorkOSClient } from '../../../../clients/workos.client.js';
 import { envs } from '../../../../env.js';
 import { asyncWrapper } from '../../../../utils/asyncWrapper.js';
